After 10. Before Optimizations

+++ backend/routes/posts.js
@@ -39,14 +39,19 @@ router.post('', checkAuth, multer({ storage: storage }).single("image"), (req, r
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({message: 'Post added successfully!', post: {
       ...createdPost,
       id: createdPost._id,
-      } });
-  });
+      }
+    });
+  })
+    .catch(error => {
+      res.status(500).json({message: 'Creating a post failed!'});
+    });
   console.log(post);
 
 });
@@ -73,9 +78,11 @@ router.get('', (req, res, next) => {
           message: 'Data fetched successfully!',
           posts: fetchedPosts,
           maxPosts: count
-        })
+        });
     })
-    .catch();
+    .catch(error => {
+      res.status(500).json({message: 'Fetching posts failed!'});
+    });
 });
 
 router.get("/:id", (req, res, next) => {
@@ -86,7 +93,10 @@ router.get("/:id", (req, res, next) => {
     } else {
       res.status(404).json({ message: "Post not found!" });
     }
-  });
+  })
+    .catch(error => {
+      res.status(500).json({message: 'Fetching a post failed!'});
+    });
 });
 
 router.put(
@@ -102,20 +112,35 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
-    });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({message: 'Post updated!'});
+      } else {
+        res.status(401).json({message: 'Not authorized!'});
+      }
+    })
+      .catch(error => {
+        res.status(500).json({message: 'Updating a post failed!'});
+      });
   }
 );
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then( result => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then( result => {
     console.log(result);
-    res.status(200).json({message: 'Post deleted!'});
-  });
+    if (result.n > 0) {
+      res.status(200).json({message: 'Post deleted!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
+  })
+    .catch(error => {
+      res.status(500).json({message: 'Deleting a post failed!'});
+    });
 });
 
 module.exports = router;
